食事の日程調整機能を追加し、Mealコンポーネントと関連するUIコンポーネントを実装
